+++ PerkUnlockHelper/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [1.2.0] - 2025-07-18
+
+### Added
+
+- Display of previously unlocked perks if they have not been acquired in this save.
+- Setting `ShowPreviouslyUnlocked` to enable displaying previously unlocked perks. Enabled by default. 
+
+### Changed
+
+- If a perk has not been unlocked and `ShowImages` is enabled, we now mute the image a bit so it can be differentiated from already unlocked perks.
+
 ## [1.1.0] - 2025-07-16
 
 ### Added
@@ -25,5 +36,6 @@ and this project adheres to [Semantic Versioning].
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
+[1.2.0]: https://github.com/LunaLycan287/StarValorMods/releases/tag/PerkUnlockHelper_v1.2.0
 [1.1.0]: https://github.com/LunaLycan287/StarValorMods/releases/tag/PerkUnlockHelper_v1.1.0
 [1.0.0]: https://github.com/LunaLycan287/StarValorMods/releases/tag/PerkUnlockHelper_v1.0.0